+++ kipper/core/src/compiler/semantics/symbol-table/universum-scope.ts
@@ -72,6 +72,7 @@ export const BuiltInFunctions = {
  */
 export const BuiltInVariables = {
 	__name__: new BuiltInVariable("__name__", BuiltInTypes.str, true),
+	NaN: new BuiltInVariable("NaN", BuiltInTypes.num, false),
 } satisfies Record<string, BuiltInVariable>;
 
 /**